Moe: Fix trunc_page to keep upper part of address
Use unsigned long for page_size such that the
expression does not cut off the upper part of the address
due to smaller size of right expression.

+++ moe/server/src/dataspace_util.cc
@@ -21,7 +21,7 @@ using Moe::Dataspace_noncont;
 
 namespace {
 
-unsigned long trunc_page(unsigned page_size, unsigned long addr)
+unsigned long trunc_page(unsigned long page_size, unsigned long addr)
 { return addr & ~(page_size-1); }
 
 inline void
